requirements.txt
pandas
openpyxl
streamlit
plotly
sentence-transformers
scikit-learn
umap-learn
numpy
scipy
